+++ .gitignore
@@ -1,14 +1,8 @@
 .DS_Store
 /node_modules
 
-# local env files
-.env.local
-.env.*.local
-
 # Log files
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
 
 # Editor directories and files
 .idea
