+++ music-for-business/styles/_altfooter.scss
