+++ package.json
@@ -1,29 +1,20 @@
 {
   "name": "steam",
   "version": "1.0.0",
-  "description": "A Gulp file and project generator.",
-  "main": "gulpfile.js",
+  "description": "",
+  "main": "",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "clean": "rm -rf public",
+    "scss": "sass music-for-business/styles/main.scss public/music-for-business/main.css -s compressed",
+    "html": "cp index.html public/index.html && cp music-for-business/index.html public/music-for-business/index.html && cp 404.html public/404.html",
+    "assets": "cp CNAME public/CNAME && cp site.webmanifest public/site.webmanifest && cp -R assets public/assets",
+    "build": "npm run clean && npm run scss && npm run html && npm run assets && rollup --config"
   },
   "author": "steam",
   "license": "ISC",
   "devDependencies": {
-    "browser-sync": "2.6.5",
-    "del": "^3.0.0",
-    "fs": "0.0.1-security",
-    "gulp": "^3.9.1",
-    "gulp-cache": "0.2.8",
-    "gulp-change": "^1.0.0",
-    "gulp-concat": "2.5.2",
-    "gulp-declare": "^0.3.0",
-    "gulp-imagemin": "2.2.1",
-    "gulp-inline-css": "^3.1.0",
-    "gulp-minify-css": "1.0.0",
-    "gulp-plumber": "1.0.0",
-    "gulp-rename": "1.2.2",
-    "gulp-replace": "^0.6.1",
-    "gulp-sass": "^2.3.2",
-    "gulp-uglify": "1.2.0"
+    "rollup": "^2.70.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "sass": "^1.49.11"
   }
 }
