+++ public/404.html
@@ -1,0 +1,56 @@
+<!doctype html>
+<html>
+  <head>
+    <style type="text/css">
+      body{
+        background: #4887f7;
+        font-family: 'consolas', sans-serif;
+        margin: 0;
+        padding: 0;
+        text-align: center;
+      }
+
+      .wrap {
+        align-items: center;
+        display: flex;
+        flex-direction: column;
+        height: 100vh;
+        justify-content: center;
+        width: 100%;
+      }
+
+      .header,
+      .subheader {
+        color: rgba(0,0,0, .4);
+        line-height: .8;
+      }
+
+      .header{
+        font-size: 160px;
+      }
+
+      .subheader{
+        font-size: 32px;
+      }
+
+      .link {
+        color: #fff;
+        display: inline-block;
+        font-size: 25px;
+        font-weight: light;
+        margin: 50px 0 0;
+      }
+
+      .link:hover {
+        text-decoration: none;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="wrap">
+      <div class="header">404</div>
+      <div class="subheader">PAGE NOT FOUND</div>
+      <div><a class="link" href="http://mobilesteamunit.com">BACK TO MOBILESTEAMUNIT.COM</a></div>
+    </div>
+  </body>
+</html>
