+++ public/music-for-business/main.js
@@ -1,0 +1,1 @@
+var t,e=document.getElementById("html5audio"),n=!0,o=!0,a={from:"15,0 50,25 50,75 15,100",to:"15,10 40,10 40,90, 15,90"},r={from:"50,25 85,50 85,50, 50,75",to:"60,10 85,10 85,90, 60,90"},i=["https://res.cloudinary.com/ddkucvkg0/video/upload/v1531767271/steam/dld-theme.mp3","https://res.cloudinary.com/ddkucvkg0/video/upload/v1531767272/steam/donut-theme.mp3","https://res.cloudinary.com/ddkucvkg0/video/upload/v1520636718/steam/guru.mp3","https://res.cloudinary.com/ddkucvkg0/video/upload/v1520636716/steam/spacious.mp3"];function c(){$("#play").on("click",d),$("#next").on("click",(function(){$("#nextanim").get(0).beginElement(),s(u()+1)})),$("#prev").on("click",(function(){$("#prevanim").get(0).beginElement(),s(u()-1)})),$(".track").on("click",(function(){s(parseInt($(this).data("trackid")))})),e.addEventListener("pause",(function(){n=!0,clearInterval(t),l()})),e.addEventListener("play",(function(){var n;n=$(".audio").outerWidth(),t=setInterval((function(){var t=e.currentTime/e.duration*n;$("#buffer").css({"-webkit-transform":"translateX("+t+"px)","-moz-transform":"translateX("+t+"px)","-ms-transform":"translateX("+t+"px)",transform:"translateX("+t+"px)"})}),100),l()})),e.addEventListener("ended",(function(){var t=u();t!==i.length-1&&s(t+1)}))}function s(t){n=e.paused,t>=i.length?t=0:t<0&&(t=i.length-1),$("[data-curtrack]").attr("data-curtrack",t),$(".track").removeClass("active"),$('[data-trackid="'+t+'"]').addClass("active"),e.src=i[t],e.load(),e.play()}function d(){f()&&o?s(0):f()?e.play():e.pause(),o=!1}function l(){n&&($("#left").attr({from:f()?a.to:a.from,to:f()?a.from:a.to}).get(0).beginElement(),$("#right").attr({from:f()?r.to:r.from,to:f()?r.from:r.to}).get(0).beginElement())}function u(){var t=$("[data-curtrack]").attr("data-curtrack");return parseInt(t)}function f(){return e.paused}var m,h={sectionOne:0,sectionTwo:0,sectionThree:0};function p(){$(window).scroll((function(){var t,e,n,o;0===h.sectionOne&&g(),t=$(window).scrollTop(),e=t>h.sectionOne&&t<=h.sectionTwo,n=t>h.sectionTwo&&t<=h.sectionThree,o=t>h.sectionThree,e?v("#fff"):n?v("#4586ff"):o&&v("#16135a")})),$(window).resize((function(){g()})),$(".trigger-contact").click((function(){var t=$(this).data("selectval");$("html, body").animate({scrollTop:$("#contact-form").offset().top-50},800),$("#request").val(t),$("#name").focus()})),$(".yt-async").click((function(){var t,e;t=this,(e=document.createElement("iframe")).setAttribute("src","https://www.youtube.com/embed/"+t.dataset.video+t.dataset.params),e.style.width=315,e.style.height=560,t.parentNode.replaceChild(e,t)}))}function v(t){$(".products").css("background",t),$(".products .product-button").css("color",t)}function g(){h.sectionOne=$(".jingles").offset().top-ScreenData.height/3,h.sectionTwo=$(".gigs").offset().top-ScreenData.height/3,h.sectionThree=$(".art").offset().top-ScreenData.height/3}function w(){ScreenData.isSmall&&$(window).scroll((function(){!function(){for(var t=0;t<m.length;t++){var e=m[t];!$(e).hasClass("hovered")&&e.getBoundingClientRect().top<k()?$(e).addClass("hovered"):$(e).hasClass("hovered")&&e.getBoundingClientRect().top>=k()&&$(e).removeClass("hovered")}}()}))}function k(){return ScreenData.height/4}window.addEventListener("load",(()=>{ScreenData={height:window.innerHeight,width:window.innerWidth,isSmall:()=>ScreenData.width<768},window.onresize=()=>{ScreenData.height=window.innerHeight,ScreenData.width=window.innerWidth},m=document.querySelectorAll(".member"),w(),$(window).resize((function(){w()})),p(),c(),document.getElementById("year").textContent=(new Date).getFullYear()}));
