--- src/styles/_footer.scss
@@ -1,180 +1,0 @@
-.footer {
-	background-color: #eee;
-	border-top: 5px solid #e2e2e2;
-}
-
-.footer__content {
-	align-items: center;
-	display: flex;
-}
-
-.footer__logo {
-	width: 200px;
-
-	.logo {
-		width: 100%;
-		opacity: .4;
-		filter: gray; /* IE6-9 */
-		-webkit-filter: grayscale(1); /* Google Chrome, Safari 6+ & Opera 15+ */
-		filter: grayscale(1); /* Microsoft Edge and Firefox 35+ */
-	}	
-}
-
-.footer__links {
-	box-sizing: border-box;
-
-	// .link-group__header {
-	// 	color: #888;
-	// 	@include heading-font;
-	// 	text-transform: uppercase;
-	// }
-
-	.footer__link {
-		color: #333;
-		cursor: pointer;
-		display: inline-block;
-		line-height: 1.5;
-		@include body-font;
-	}
-
-}
-
-.footer__extras {
-	color: #888;
-	max-width: 430px;
-	@include body-font;
-
-	.extras__signoff {
-		display: inline-block;
-	}
-
-	.extras__copyright {
-		margin-top: 20px;
-
-		.mm-link {
-			text-decoration: none;
-
-			.mm-logo {
-				fill: #333;
-				height: 20px;
-				width: 20px;
-				@include transition(200ms);
-			}
-		}
-	}
-}
-
-
-@media (min-width: 768px) {
-	.footer {
-		padding: 80px;
-
-		.footer__content {
-			flex-direction: row;
-		}
-
-		.footer__links {
-			flex: auto;
-			margin: 0 0 0 50px;
-			padding: 0 30px;
-
-			.footer__link-group {
-				float: left;
-				margin: 0 50px 0 0;
-			}
-
-			.footer__link {
-				@include transition(150ms);
-
-				&:hover {
-					color: $blue;
-					@include transform(translateX(2px));
-				}
-			}
-		}
-
-		.footer__extras {
-			text-align: right;
-
-			.extras__signoff {
-			}
-
-			.extras__copyright {
-				height: 30px;
-				line-height: 30px;
-				vertical-align: middle;
-
-				.mm-link {
-					cursor: pointer;
-					display: inline-block;
-					height: 30px;
-					line-height: 30px;
-					text-decoration: none;
-					vertical-align: middle;
-
-					&:hover .mm-logo {
-						fill: $blue;
-						@include transform(rotate(360deg));
-					}
-
-					.mm-logo {
-						height: 30px;
-						width: 30px;
-					}
-				}
-			}
-		}
-	}
-}
-
-@media (max-width: 767px) {
-	.footer {
-		padding: 30px;
-
-		.footer__content {
-			flex-direction: column;
-		}
-
-		.footer__logo {
-			margin: 0 0 25px;
-		}
-
-		.footer__links {
-			display: flex;
-			flex-direction: row;
-			width: 100%;
-
-			.footer__link-group {
-				flex: auto;
-				text-align: center;
-
-				.footer__link {
-					line-height: 2.5;
-				}
-			}
-		}
-
-		.footer__extras {
-			line-height: 1.4;
-			margin: 0 auto;
-			padding: 20px 0 0;
-			text-align: center;
-
-			.extras__signoff {
-				max-width: 80%;
-			}
-
-			.extras__copyright {
-				.mm-link {
-					display: block;
-					margin: 15px 15px 0;
-
-					.mm-logo {
-						height: 50px;
-						width: 50px;
-					}
-				}
-			}
-		}
-	}
-}
