--- src/styles/_header.scss
@@ -1,138 +1,0 @@
-.hero {
-  --heading-font: 'Patua One', Verdana, sans-serif;
-  --body-font: 'Source Sans Pro', Helvetica, Arial, sans-serif;
-  --left-margin: 15%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  overflow: hidden;
-  position: relative;
-  
-  &::before {
-    background-image: url('https://res.cloudinary.com/ddkucvkg0/image/upload/v1531848110/steam/culture-first-warp2.jpg');
-    background-size: cover;
-    bottom: 0;
-    content: "";
-    filter: contrast(115%);
-    left: 0;
-    position: absolute;
-    right: 0;
-    top: 0;
-    z-index: 0;
-  }
-
-  .hero__logo {
-    width: 160px;
-    z-index: 1;
-  }
-  
-  .hero__headline,
-  .hero__subheadline {
-    color: #fff;
-    flex: none;
-    position: relative;
-    z-index: 1;
-  }
-
-  .tm {
-    font-family: Arial, Helvetica, sans-serif;
-    font-size: 18px;
-    line-height: 40px;
-    vertical-align: text-top;
-  }
-
-  .hero__cta {
-    background-color: #2ac971;
-    border-radius: 6px;
-    box-shadow: 0px 7px 10px rgba(0,0,0,.1);
-    color: #fff;
-    cursor: pointer;
-    font-family: var(--body-font);
-    font-size: 20px;
-    font-weight: 900;
-    max-width: 300px;
-    padding: 15px 0;
-    position: relative;
-    text-shadow: 0px 0px 1px rgba(0,0,0,.6);
-    text-align: center;
-    text-transform: uppercase;
-    transition: 175ms;
-    width: 300px;
-    -webkit-font-smoothing: antialiased;
-    z-index: 1;
-
-		&:hover {
-			box-shadow: 0px 13px 15px rgba(0,0,0,.1);
-			letter-spacing: 1px;
-			transform: translateY(-2px);
-		}
-  }
-}
-
-
-@media (min-width: 768px) {  
-  .hero {
-    height: 100vh;
-
-    &::before {
-      background-position: center;
-    }
-
-    .hero__headline,
-    .hero__subheadline,
-    .hero__cta {
-      margin-left: 15%;
-    }
-
-    .hero__logo {
-      left: 30px;
-      position: absolute;
-      top: 30px;
-    }
-
-    .hero__headline {
-      font-size: 65px;
-      font-family: var(--heading-font);
-    }
-
-    .hero__subheadline {
-      font-size: 36px;
-      font-family: var(--body-font);
-      line-height: 1.3;
-      margin: 10px 0 25px var(--left-margin);
-      max-width: 500px;
-    }
-  }
-}
-
-@media (max-width: 767px) {
-  .hero {
-    height: auto;
-    text-align: center;
-
-    &::before {
-      background-position: right;
-    }
-
-    .hero__logo {
-      margin: 50px auto 0;
-      position: relative;
-    }
-
-    .hero__headline {
-      font-size: 30px;
-      margin: 60px 0 30px;
-    }
-
-    .hero__subheadline {
-      font-family: var(--body-font);
-      font-size: 26px;
-      margin: 0 25px 50px;
-    }
-
-    .hero__cta {
-      margin: 0 auto 50px;
-    }
-  }
-}
