--- src/styles/critical.scss
@@ -1,2 +1,0 @@
-@import 'reset';
-@import 'header';
